requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
tortoise-orm==0.21.3
aerich==0.7.2
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.8.0
pydantic-settings==2.4.0
python-jose==3.3.0
httpx==0.27.0
python-multipart==0.0.9
boto3==1.35.0
Pillow==10.4.0
anthropic==0.34.0
stripe==10.9.0
PyMuPDF==1.24.0
slowapi==0.1.9
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
aiosmtplib==3.0.1
python-jose[cryptography]==3.3.0
email-validator==2.3.0
pytest==8.3.0
pytest-asyncio==0.23.8
httpx==0.27.0